requirements.txt
fastapi
passlib
pyjwt
python-multipart
sqlmodel
uvicorn
pymysql
passlib[bcrypt]
